2025-2029/Adithyanandan/NeuralFix/Web-App/App/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
alembic>=1.18.4
pydantic>=2.10.0
pydantic-settings>=2.6.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0
Pillow>=11.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
ollama>=0.3.0
groq>=0.11.0